[chore][pkg/stanza] Cleanup secondary parser operator files
