+++ pkg/stanza/operator/parser/scope/config.go
@@ -4,11 +4,8 @@
 package scope // import "github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/operator/parser/scope"
 
 import (
-	"context"
-
 	"go.uber.org/zap"
 
-	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/entry"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/operator"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/operator/helper"
 )
@@ -49,15 +46,4 @@ func (c Config) Build(logger *zap.SugaredLogger) (operator.Operator, error) {
 		TransformerOperator: transformerOperator,
 		ScopeNameParser:     c.ScopeNameParser,
 	}, nil
-}
-
-// Parser is an operator that parses logger name from a field to an entry.
-type Parser struct {
-	helper.TransformerOperator
-	helper.ScopeNameParser
-}
-
-// Process will parse logger name from an entry.
-func (p *Parser) Process(ctx context.Context, entry *entry.Entry) error {
-	return p.ProcessWith(ctx, entry, p.Parse)
 }
