+++ pkg/stanza/operator/parser/scope/parser_test.go
