+++ pkg/stanza/operator/parser/severity/parser_test.go
