+++ pkg/stanza/operator/parser/time/parser_test.go
