+++ pkg/stanza/operator/parser/trace/config.go
@@ -4,11 +4,8 @@
 package trace // import "github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/operator/parser/trace"
 
 import (
-	"context"
-
 	"go.uber.org/zap"
 
-	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/entry"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/operator"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/operator/helper"
 )
@@ -53,15 +50,4 @@ func (c Config) Build(logger *zap.SugaredLogger) (operator.Operator, error) {
 		TransformerOperator: transformerOperator,
 		TraceParser:         c.TraceParser,
 	}, nil
-}
-
-// Config is an operator that parses traces from fields to an entry.
-type Parser struct {
-	helper.TransformerOperator
-	helper.TraceParser
-}
-
-// Process will parse traces from an entry.
-func (p *Parser) Process(ctx context.Context, entry *entry.Entry) error {
-	return p.ProcessWith(ctx, entry, p.Parse)
 }
