+++ pkg/stanza/operator/parser/trace/parser_test.go
